vmbase: x86_64: add idmap and 32-to-64-bit mode switch
Add a minimal identity-mapped page table for x86-64 pvmfw, and set up
paging using this idmap in the vmbase entry point.

Additionally, detect if the machine is currently running in 32-bit mode
at the entry point and switch to 64-bit long mode if so. 32-bit mode
will be used for the pVM firmware entry point in the future, but the
current pKVM and crosvm implementation enters protected VMs in 64-bit
long mode, so support both modes for now.

Bug: 393356668
Test: m pvmfw
Flag: EXEMPT b/393356668
Change-Id: Iac350c85a329b435d4bfa46b5d5716807bd7d6b0