virtmgr: sleep soundly in monitor_vm_status
The thread was waking up every second, but only doing work every 30
seconds. It was probably done that way to enable reacting to the VM
dying. Instead, use a condition variable so that the thread can be
immediately notified of VM death and sleep for the full 30 seconds.

Flag: EXEMPT ...
Test: boot VM, take trace, see virtmgr thread wake up less often
Change-Id: I60a5aea7605d48d958f8de0dd2941c57ed99a032