microdroid: disable first_stage_console
This only affects "debuggable" builds because `init` disables the
`first_stage_console` feature at compile time for `-user` builds.

When enabled, it causes `init` to do some extra work and then execute
`/first_stage.sh`, which doesn't exist in microdroid, and then hook up a
shell to the console. Based on the logs, it can take 4ms on an idle
device, so it is a minor win to disable it.

    [    0.307263][   T40] init: Attempting to run /first_stage.sh...
    [    0.308481][   T41] init (41) used greatest stack depth: 14208 bytes left
    [    0.309322][   T40] init: /first_stage.sh exited
    [    0.310315][   T40] init (40) used greatest stack depth: 13680 bytes left
    [    0.311180][    T1] init: console shell exited

It can always be enabled via the cmdline if needed for local debugging.
Note that this feature looks like a backdoor at first glance, it gives
host access to a root shell inside pVMs and it was enabled even in the
non-debuggable bootconfig, but, as mentioned above, it is disabled in
release builds via a compile time param.

Bug: 412472524
Test: boot microdroid
Flag: EXEMPT ...
Change-Id: I1af1412a8fa63a38069af43626fda1dd2ce7967f