security_vm: Add binder accessor for keymint ICommService in VM
This cl enables the security_vm_launcher to expose a binder
accessor running inside the Trusty Security VM to the host.

When the launcher is started with a `--port` argument, it creates
and registers an accessor binder. This accessor bridges the host
to the RPC service in the VM.

Bug: 404289721
Test: launch KeyMint connecting to trusty VM
Test: atest VtsAidlSharedSecretTargetTest
Flag: EXEMPT this change only affects Trusty qemu
Change-Id: I090ad39ac4b174cab7b2b3d56554d2350e8b17be