pvmfw: config: No panic if missing DICE handover
A previous change added a "hidden" ABI for QTVMs that would behave
differently if the DICE handover was missing. However, the low-level
parser contained a check that would cause it to panic. Therefore,
remove the check. This wasn't caught in CI because we currently have no
test coverage for this code path.

Fixes: ac1a420569f6 ("pvmfw: Support QTVMs through missing DICE handover")
Flag: EXEMPT b/406977494
Bug: 393977894
Test: -
Change-Id: If19312cf08e0539db79824b3b01d1db96a8f3f58