build/debian: Split off kernel headers and modules from rootfs
By default, we currently boot the Ferrochrome kernel directly,
which lives outside of the roofs.
However, the kernel DTBs, headers and modules all remain in the rootfs.

In this CL, we move these out of rootfs into a new disk image,
which uses a fixed UUID that also appears in the `/etc/fstab`.
This allows upgrading kernel-related files without touching the rootfs.
Furthermore, we attach and mount this new disk as read-only,
thereby preventing potential overwrites during system upgrades.
Symbolic links are used to map the rootfs directories for
DTBs, headers and modules into this new image.

Finally, to clean up the rootfs,
we purge the kernel image, modules, and headers related packages.

Bug: 409484190
Flag: EXEMPT refactor
Test: ./build_in_container.sh -a aarch64 ; then deployed on caiman

Change-Id: I0e0a964da46c7455747d187b1c399521d6c62247