vrtmgr: log EOF for VM shutting down
VMs shutting down is an expensive operation,
so one log doesn't hurt.

In b/404210068, I was thinking perhaps we are
missing the last few logs from the VM, when
this process gets shutdown. So, add more info,
so we can more easily see whether we properly
reach the end of stream.

Bug: 404210068
Test: N/A
Change-Id: Icf1dfc90b6608e87fe9cfb5042c887d40cf5f6ac