Persist encrypted store mode on disk in config.xml
This change adds a concept of encrypted store represented as @IntDef
enum to the VirtualMachineConfig and persists it on disk as part of the
config.xml.

Currently there are 2 modes:
* DEFAULT - this mode is used by most of the VMs.
* ENCRYPTED_STORE_KEK_ON_CE - this mode is going to be used by the
  appsearch VM. It allows starting a VM before a user is unlocked, while
  having the encrypted store of this VM still being protected by the
  secret only available after user unlocks the device.

We are adding this concept into AVF stack instead of the appsearch to
make future improvements easier. Additionally, persisting the mode on
disk allows us to implement the migration from "DEFAULT encrypted store
on DE" mode to the new "encrypted store is located on DE, it's KEK is
located on CE" mode.

Bug: 406258175
Test: atest MicrodroidTests
Flag: build.RELEASE_AVF_SUPPORT_LONG_RUNNING_VMS
Change-Id: Ifeb51a17edd15a6651db2c801d1e105fca469f9d