Parse ENCRYPTED_STORE_MODE property in microdroid
This information will be used in microdroid_manager to decide which
encrypted store mode to use. Since the property is defined in the
AndroidManifest.xml of the main apk, microdroid_manager can trust it.

Since I was there, refactored the delayed encrypted store setup logic
into a separate function. In the follow-up patch this function will be
extended to support new mode.

Bug: 406258175
Test: atest MicrodroidTests
Flag: build.RELEASE_AVF_SUPPORT_LONG_RUNNING_VMS
Change-Id: Iacc956ac10c244ac2878197c996d50373bda11c7