pageout_bomb: new command to MADV_PAGEOUT all processes in the guest
This is intended to be an alternative to inflating the balloon. The main
side effect of inflating the balloon to max is that it drops caches in
the kernel and forces userspace memory into swap. The theory here is
that it would be better to do those things directly, with the advantage
that there is less risk of causing OOMs (because MemAvailable doesn't go
down) and then let VIRTIO_BALLOON_F_PAGE_REPORTING release the pages to
the guest proactively.

Using `setprop pageout_bomb.go 1` (invokes pageout_bomb via init rc
files) gives worse results than shelling into the VM and invoking
`/system/bin/payout_bomb` directly. It may be because `init` touches a
fair amount of pages as part of reaping the process.

Bug: 418304368
Test: launch microdroid VM, connect, `setprop pageout_bomb.go 1`, see memory usage go down
Flag: EXEMPT ...
Change-Id: Iab7a1d51f6a01e3b58772251c405e62a96280887