graceful shutdown in microdroid
With this change, AVF tries to shut virtual machines down gracefully.
Specifically, when a VM is requested to be stopped, it first checks if a
guest agent was registered from the VM. If so, it sends a command to
initiate a shutdown sequence. If the complete shutdown of the VM is not
observed within 5 seconds, it then falls back to killing the VM
forcibly.

Currently, Microdroid is the only guest OS which implements the guest
agent. The guest agent initiates the shutdown sequence by setting
sys.shutdown.requested sysprop. Then the app payload is expected to
handle the event (by monitoring any change to the sysprop). At the same
time, init starts a shutdown watchdog sequence where it tries to sync
disk accesses as much as possible. This has to be done within 2 seconds,
and after that the guest-side kernel shuts the power down by itself.

Bug: 419607474
Flag: EXEMPT bugfix
Test: android/vm/vm_shell.sh start-microdroid --auto-connect
kill `vm` process in another adb shell terminal
guest-side log:

[   15.904284][   T43] init: Starting RebootMonitorThread
[   15.905895][   T87] init: RebootMonitorThread started for 2s
[   16.108599][   T87] init: Reboot monitor still running, forced reboot in 1800 ms
[   16.110816][   T87] sysrq: Emergency Sync
[   16.112230][    T8] Emergency Sync complete
...
[   18.088347][   T87] init: Show tasks that are in disk sleep(uninterruptable sleep), which are like blocked in mutex or hardware register access:
[   18.093206][   T87] sysrq: Show Blocked State
[   18.094075][   T87] init: Reboot ending, jumping to kernel
[   18.096462][   T87] reboot: Power down

Change-Id: I74cce63382ac4a4002fc4a5958315718c6f068f5