Report DICE policy errors with the PAYLOAD_CHANGED error code
When a DICE policy error occurs, it indicates that the payload has
changed in an incompatible way. That might mean the version has
downgraded, or more/fewer components are present, etc but the result is
that the VM will not be able to access its previous secrets. This is an
event that the client app needs to notice and handle so it needs to be
exposed better than an UNKNOWN error.

There's already a PAYLOAD_CHANGED that was used when strict version
binding was the only option and no upgrade of VM payloads was allowed.
This has become unused for updatable VMs but feels appropriate to make
use of again when the change to the payload was illegal, as identified
by the DICE policy error.

Bug: 420367677
Test: atest MicrodroidTests
Flag: EXEMPT bugfix
Change-Id: Icc09930277de0911528c5fd4a9c640623553125a