Keep virtmgr in the foreground process group
With ag/33580636, virtmgr and crosvm run in a new process group. As a
side effect of this, they are detached from the controlling terminal.
Then when they write something to stdout, they are stopped via SIGTTOU.

To fix this, this change uses tcsetpgpr to make the new process group
foreground.

Bug: 422532266
Test: adb shell and then run `vm run-microdroid`
Test: adb shell and then run `echo "" | vm run-microdroid`
Flag: EXEMPT bug fix
Change-Id: Id67b139400f22e2dcd3b635ba74b6dc2755f1269