Restart psi_monitor on failure and try on EINTR
Linux can EINTR epoll.wait, which was causing the psi_monitor thread to
return an Error resulting in the VM shutting down.

Instead, check if the error is EINTR and wait again.

If the psi_monitor thread fails, start it again, until the VM is asked
ot shut down resulting in the epoll receiving the kill event.

Bug: b/424302767
Test: boot VM, interrupt epoll thread.
Flag: EXEMPT ...
Change-Id: I46eea96a7a3b989b630d089f3619d3c9b2f781f2