VM is registered to the global service
When a new IVirtualMachine object is created, it is registered to the
global service (virtualizationservice), so that we can collect the list
of all virtual machines on the system via the global service.

Furthermore, if the global service is killed, virtmgr re-starts it (via
death recipient), and re-registers the VMs it manages to the service, so
that the list of VMs the global service maintains is always up to date.

Note that this change doesn't yet make use of the registered VMs.
Follow-up changes will do.

Bug: 418877672
Flag: EXEMPT bug fix
Test: run two VMs, kill one by one, observe logcat
I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2048 registered
I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2049 registered
I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2048 unregistered
I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2049 unregistered

Test: run two VMs, kill virtualizationservice, observe logcat. Notice
the difference in PID.
2727  2732 I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2050 registered
2727  2732 I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2051 registered
2848  2851 I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2051 registered
2848  2854 I VirtualizationService: virtualizationservice::aidl: Virtual machine with CID 2050 registered

Change-Id: I88696e9d000d07b7d7ddb9b04775ec8d4f720458