Dereference VmInstance from Strong<dyn IVirtualMachine>
The relationship between the objects around VirtualMachine is as follows

Strong<dyn IVirtualMachine>   // binder pointer
    |
    |(Box)
    v
Binder<BnVirtualMachine>    // binder object
    |
    |(Box)
    v
VirtualMachine  // user object
    |
    |(Arc)
    v
VmInstance  // user object

So, if you have Strong<dyn IVirtualMachine>, you don't have to demand a
reference to VmInstance, because the latter can be dereferenced from the
former.

This change is in preparation for the merging of VirtualMachine and
VmInstance.

Bug: 424687020
Test: N/A
Flag: EXEMPT refactoring
Change-Id: Ibb858df92f3fb69bf6ae46b70d10a8f094e13f5b