Ensure that VirtualMachineCallback is for the current VM
For a given VirtualMachine object, only one VirtualMachineCallback can
be registered at a time; `VirtualMachine.setCallback`. This means that
if you register one callback and then re-register a new callback, the
first callback shouldn't be notified.

However, there was a bug that the first callback can get notified. This
happens when you re-start a VM with a new callback without waiting
for the `onStopped` event through the first callback. Then by the
asychronous nature of VM lifecycle management mechanism, the stopping of
the prior VM can get notified to the second callback which would cause
confusion as it's not clear whether it's for the first run or the second
run.

This bug gets manifested as multiple instances of the same VM.
Specifically, it happens in the following steps:

1. a VM is started with a callback.
2. that VM somehow gets killed. (due to OOM, or whatever)
3. the owner of the VM detects this event via DEAD_OBJECT exception, or
   DeathRecipient if there was a binder connection to the VM.
4. the owner of the VM re-starts the VM, with a new callback CB2.
5. While the second VM gets started, the death of the first VM is
   notified via the new callback.
6. upon receiving onStopped, the owner of the VM, re-starts the VM.
7. We now have two VM instances.

This CL fixes this issue by ensuring that VirtualMachineCallback is used
only for the current VM. It does this by comparing the CID of the
callback event and the CID of the current virtual machine.

Bug: 428038058
Test: atest MicrodroidTestApp:com.android.microdroid.test.MicrodroidTests#stopAndRestart
(confirmed that the test fails without the fix)
Flag: EXEMPT bug fix

Change-Id: I4a5e6f2f518014bd3f0ad893ce079ce8ca1cf38b