Drop Debug from CrosvmConfig and CrosvmCommand
They were Debug just because VmState which contains CrosvmConfig wanted
to be Debug. However, CrosvmConfig and CrosvmCommand may want to add
non-debuggable fields in the future (actually they are), which currently
is impossible.

VmState is Debug only because we just want to know whether a VM is
running, stopped, etc., not because we want to know the internal details
of CrosvmConfig and CrosvmCommand.

So, let's implement the Debug trait for VmState manually, and drop the
trait from the Crosvm* structs.

Bug: N/A
Test: N/A
Flag: EXEMPT refactoring
Change-Id: I3130d3f4233f5e49f7e8d41d2932e48ddc2cb4af