pvmfw: Fix panic on missing 'interrupts' in VM DT
Not all assignable nodes in the VM device tree may have an 'interrupts'
property.
If a malicious host adds this property via the VMM, attempting to remove
the non-existent 'interrupts' property from the VM DT can result in an
error.

This patch masks that error, as it's harmless.
It ensures that the final device tree reflects only what the VM DTBO
specifies, ignoring untrusted modifications from the host.

Test: Booted QTVM with pvmfw on Qualcomm SM8850 chipset.
Flag: EXEMPT b/406977494
Fixes: ebf4488e846c ("pvmfw: Do not return error when interrupts property is not present")
Bug: 385802423
Bug: 422213097
Reviewed-on: https://partner-android-review.googlesource.com/3344578
Change-Id: I250513997d085aa35a5e69fc552cc58946d6fc8d
Signed-off-by: Sreenad Menon <quic_sreemeno@quicinc.com>