Stop all VMs during shutdown
Previously, virtualizationservice waited on sys.powerctl to detect the
start of shutdown. Howwever, sys.powerctl is a special sysprop which
cannot be read by processes other than init. So it actually wasn't
working.

In this change, we instead wait for the SIGTERM signal. init sends the
signal very early in the shutdown process. Upon receiving the signal,
virtualizationservice stops all VMs in the system.

Bug: 430518704
Flag: EXEMPT bugfix
Test: run a VM, setprop sys.powerctl reboot. Check logcat to see if
"SIGTERM received. Stopping VMs..." is shown.

Change-Id: I82de6d1e08540610d75a275ec07221b288950e80