pvmfw: allow /memory to include the pvmfw region
When combined with a crosvm change to put the pvmfw region in /memory,
this allows the guest to reuse the 4 MB initially reserved for pvmfw
(minus the 4 KB reserved region for DICE).

This effectively increases the guest RAM size by ~4 MB for pVMs. We may
want to compensate by subtracting 4 MB off the requested RAM size, but
it is tricky for small VMs, e.g. if a client requests only 1 MB, we
can't reduce it to 0 MB because we need space outside the pvmfw region
to copy the image into.

Bug: 431065184
Test: boot microdroid
Flag: EXEMPT .
Change-Id: I44ad38d112d258fbe7e719bd836f84eb85be6ab7