Reland: GTS/Multi-tenant: Test devices with sdk >= 35
The corresponding feature flag was removed in b/365767108. The
side-effect of this is the tests are being exercised on older platform
versions as well, which is unexpected. Ensure the test is run only if
Build.VERSION.SDK_INT >= 35 (or check in case of
testFileUnderBinHasExecutePermission, check for correct permission
conditional to platform support)

Test: Test builds
Flag: TEST_ONLY
Bug: 429485472
Bug: 432419807
Change-Id: I1c71616a101eda745f7282455259dbdd135ca8fc