sign_virt_apex: Relax assert, to pass GTS
Since 199c20d5eaae ("pvmfw: Use com.android.virt.name for RKP VM")
modified the number of AVB properties that the RKP VM ("Rialto") image
has, the assertion needed to be updated. However, as the script was
being used to resign images during GTS, the test fails as these ABIs are
not expected to be stable.

As a stopgap until the use of those scripts in GTS (or validity of those
assertions in that context) is figured out, relax the assertions so that
they pass with or without commit 199c20d5eaae.

Test: TH
Bug: 432403626
Bug: 433265999
Flag: EXEMPT GTS fix
Change-Id: Ie29f6931c1a0aab54532649770c86cf11aba960e