microdroid_mgr/swap: Only enable desired compression algorithm if available
Commit 04fe09e81f07 ("microdroid_manager: swap: Add support for
selecting a compression algo") blindly attempts to select a compression
algorithm. This is not ideal, as the kernel in use may not support the
desired compression algorithm, in which case microdroid_manager will
fail to enable swap usage, and the system will refuse to boot.

Not supporting the desired compression algorithm is not the end of the
world. The system should be allowed to run without the specific
algorithm; it just may not perform as expected. Therefore, make the
compression algorihm selection best-effort (i.e. enable the desired
algorithm only if it's available).

Bug: 433840381
Flag: EXEMPT bugfix
Fixes: 04fe09e81f07 ("microdroid_manager: swap: Add support for selecting a compression algo")
Test: MicrodroidHostTestCases com.android.microdroid.test.MicrodroidHostTests#testDebuggerdBacktrace_protectedVm_false_os_microdroid
Change-Id: I7ea554f32b0690dda0b2d15033375ad31a0f245a
Signed-off-by: Isaac J. Manjarres <isaacmanjarres@google.com>