microdroid: Reduce TCP and UDP hash table sizes
The kernel currently allocates a non-negligible amount of memory for TCP
and UDP hash tables. However, microdroid does not plan to support a
networking stack; the microdroid kernel only has networking support
because SELinux needs it.

Therefore, tune the size of these hash tables to be as small as
possible.

Bug: 434256355
Flag: EXEMPT ...
Change-Id: I60657120fe66e116e27ddf5e09cec63f39ab0a3d