virtmgr: Use rodata=on kernel command line parameter for appsearch VM
The rodata feature allows the linear mapping to be updated to read-only
to match the virtual mapping to eliminate the possibility of the data
being overwritten. However, this comes at the cost of having to map RAM
down to the base page size, instead of using block mappings, which
results in additional memory usage for pagetables. Additionally, the
number of pagetable entries can cause additional TLB pressure and
degrade performance.

Use rodata=on for the appsearch VM, since it protects the kernel image's
read-only section from being written to, and has most of the security
benefits as using rodata=full. It also means that we don't have to map
RAM to the base page size, which means we can save memory.

Bug: 437196556
Flag: EXEMPT ...
Change-Id: I17b1e9969230baab933818ea12a0d6b710a8ee28