VirtualMachine.setConfig: atomically replace config file
To ensure there is always a valid config on disk, write out the config
to a temporary path, then move it to the final location (implicitly
unlinking the old file).

Before, if config serialization failed somehow, we would end up with a
config file on disk that is either missing entirely, or invalid (e.g.
empty). Once that happens it becomes impossible to start the VM or
repair it using the existing APIs. The only option is to delete the VM
and suffer data loss.

Bug: 437160991
Test: throw exception in VirtualMachineConfig.serialize, check existing VM's on disk config doesn't become empty
Flag: EXEMPT .
Change-Id: I808a6f5f94d82b2a3b410b7a2fcf144bdc61f742