Pass the hostname= to the Microdroid kernel cmdline
Use the name in the VM config as value for the hostname= arg.
Additionally validate the value of the hostname= arg in pvmfw.

Since the value is passed from the Android host, the code inside the
protected VM shouldn't really make a lot of assumptions on it. However,
the hostname can be useful in some cases, e.g. if we want to specify a
a tracing config applicable only for the given VM.

Bug: 432805928
Bug: 222347044
Test: atest MicrodroidTests
Flag: EXEMPT, simple change
Change-Id: I96720286c43f90e9d3ce4763813967dfd148f063