VirtualMachine: throw exception on config load failure
This reverts to the pre I80e530132a0e714d6bf868d175e04b6a69ba0ddc
behavior of throwing an exception.

If `load` returns null, it is interpretted as "VM does not exist" and
client will naturally react by calling `create`. However, the `create`
code thinks the VM exists if the VM dir is present, even if the config
file is broken, so it will throw a "VM already exists" exception. The
two methods disagreeing about the state of the VM means APIs like
`getOrCreate` don't work.

For AVF to automatically recover for a bad config file, it probably
needs to delete the VM dir. That is a bigger change, so just revert to
old behavior for now.

Bug: 441586847
Bug: 437160991
Flag: NONE .
Test: manually delete VM config file on disk, reboot phone, check if VM use case can successfuly recover
Change-Id: I077798f33da9d788221b8fcbadce992bf35aebea